Add files via upload
I had a go at adding additional and advanced DOM manipulation to the stripe div. I wanted to change the selected class to display the picked color and toggle back to the regular selected class upon a restart, but am having trouble with a few bugs. Will hopefully find a solution soon.