Add files via upload
2 main updates:

The selected mode button now changes to be the picked color when chosen. Bug remains when winning round and switching modes to easy, is fixed when resetting or switching back and forth between modes

Added score tracker feature, however is not complete until I figure out how to terminate the event listener upon selecting the correct color. As long as you don't cheat, it still works!